Move notify message building into Redash class
